[spark] Move files to battery_analyzer Directory
