+++ spark/battery_analyzer/src/main/java/com/revature/ml/BatterySuccessPredictor.java
